api/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector==2.2.9
pandas
xgboost
